+++ webui/static/webui/js/edit.js
@@ -349,12 +349,7 @@
       const totalSeconds = (hours * 3600) + (minutes * 60) + seconds + (milliseconds / 1000);
 
       if(isNaN(totalSeconds)) {
-         if(returnNull) {
-            return null;
-         }
-         else {
-            return 0;
-         }
+         return returnNull ? null : 0;
       }
 
       return totalSeconds;