+++ webui/templatetags/filters.py
@@ -1,10 +1,17 @@
 from django import template
 
-from webui.utils import format_seconds
+from ..utils import format_seconds, get_version
 
 
 register = template.Library()
 
+# Function: seconds_to_segment_time
 @register.filter(name='seconds_to_segment_time')
 def seconds_to_segment_time(value):
    return format_seconds(value, segment_time=True)
+
+
+# Function: version
+@register.filter(name='version')
+def version(prefix = ''):
+   return get_version(prefix)